+++ packages/generator-adp/src/app/questions/helper/choices.ts
@@ -1,5 +1,6 @@
-import { AppRouterType } from '@sap-ux/adp-tooling';
-import type { CFApp, SourceApplication } from '@sap-ux/adp-tooling';
+import { AppRouterType, getEndpointNames } from '@sap-ux/adp-tooling';
+import type { CFApp, Endpoint, SourceApplication } from '@sap-ux/adp-tooling';
+import type { AdaptationDescriptor } from '@sap-ux/axios-extension';
 
 interface Choice {
     name: string;
@@ -60,4 +61,39 @@ export const getAppRouterChoices = (isInternalUsage: boolean): { name: AppRouter
         });
     }
     return options;
+};
+
+/**
+ * Returns the choices for the adaptation prompt.
+ *
+ * @param {AdaptationDescriptor[]} adaptations - The adaptations to get the choices for.
+ * @returns {Array<{ name: string; value: AdaptationDescriptor }>} The choices for the adaptation prompt.
+ */
+export const getAdaptationChoices = (
+    adaptations: AdaptationDescriptor[]
+): Array<{ name: string; value: AdaptationDescriptor }> => {
+    return adaptations?.map((adaptation) => ({
+        name: adaptation.title ? `${adaptation.title} (${adaptation.id})` : adaptation.id,
+        value: adaptation
+    }));
+};
+
+/**
+ * Returns the choices for the system prompt.
+ *
+ * @param {string[]} systems - The systems to get the choices for.
+ * @param {string} defaultSystem - The default system.
+ * @returns {Array<{ name: string; value: string }>} The choices for the system prompt.
+ */
+export const getKeyUserSystemChoices = (
+    systems: Endpoint[],
+    defaultSystem: string
+): Array<{ name: string; value: string }> => {
+    const endpointNames = getEndpointNames(systems);
+    return endpointNames.map((name) => {
+        return {
+            name: name === defaultSystem ? `${name} (Source system)` : name,
+            value: name
+        };
+    });
 };