+++ examples/odata-cli/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/odata-cli
 
+## 0.17.33
+
+### Patch Changes
+
+- Updated dependencies [0492325]
+    - @sap-ux/axios-extension@1.25.7
+
+## 0.17.32
+
+### Patch Changes
+
+- c707af1: fix(deps): update dependency dotenv to v16.6.1
+
 ## 0.17.31
 
 ### Patch Changes